// Angular/Tutorialv2/src/app/model/product.ts
export interface Product {
  _id?: string;
  name?: string;
  price?: number;
  description?: string;
  categories?: []
}

// export class Product {
//   _id?: string;
//   name?: string;
//   price?: number;
//   description?: string;
//   categories?: []
//
//   constructor(
//     _id?: string,
//     name?: string,
//     price?: number,
//     description?: string,
//     categories?: []
//   ) {
//     this._id = _id;
//     this.name = name;
//     this.price = price;
//     this.description = description;
//     this.categories = categories;
//   }
//
// }