// Angular/Tutorialv2/src/app/service/product.service.ts
import {Injectable} from '@angular/core';
import {Product} from "../model/product";

@Injectable({
  providedIn: 'root'
})
export class ProductService {

  constructor() {
  }

  products: Product[] = [{
    id: 1,
    name: 'IPhone 12',
    price: 2400000,
    description: 'New'
  }, {
    id: 2,
    name: 'IPhone 11',
    price: 1560000,
    description: 'Like new'
  }, {
    id: 3,
    name: 'IPhone X',
    price: 968000,
    description: '97%'
  }, {
    id: 4,
    name: 'IPhone 8',
    price: 7540000,
    description: '98%'
  }, {
    id: 5,
    name: 'IPhone 11 Pro',
    price: 1895000,
    description: 'Like new'
  }];

  getAll() {
    return this.products;
  }

  saveProduct(product: Product) {
    this.products.push(product);
  }

  findProductById(id: number) {
    for (let pro of this.products) {
      if (id == pro.id) {
        return pro;
      }
    }
    return undefined;
  }

  updateById(product: Product) {
    for (let index in this.products) {
      if (product.id == this.products[index].id) {
        this.products[index] = product;
        return true;
      }
    }
    return false;
  }

  deleteById(product: Product) {
    const index = this.products.indexOf(product, 0);
    if (index > -1) {
      this.products.splice(index, 1);
      return true;
    } else {
      return false;
    }
  }

}
